Remove handling of overflow
We've tried two different approaches for handling overflows: Handle it
in during arithmetic operations, and handle it after having solved the
constraint system.

Both solutions have distinct issues: The first can break the narrowing
step, because overflow handling hides the actual growth patterns of
variables. Handling it afterwards can needlessly widen intervals such as
[0, inf] to [-inf, inf].

We will need to revisit overflow handling once we make use of VRP to
detect impossible code branches, but the current implementation was not
the solution.